pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "openai>=1.107.1",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "qdrant-client[fastembed]>=1.14.2",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
]
